requirements.txt
asn1crypto==0.24.0
cachetools==2.0.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.4
google-auth==1.2.1
google-auth-oauthlib==0.2.0
google-oauth==1.0.1
gspread==0.6.2
idna==2.6
oauthlib==2.0.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==17.5.0
python-memcached==1.59
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
urllib3==1.22
